cleaned up removed preload auton path calls/commands
Added speed check to intake command in case the positions do not match

reverted indexing/safeToMove/havePiece to use @getter again

rewrote side auto for the 100th time. Please let this be the last. It won't, I'm too much of a perfectionist

created side 2 autoon groups with placeholder algae paths. Need to modify in the future

there is possibly an issue if the coral is intaked but not shot it will be shot at L1. This could cause major issues. We need to test